pyproject.toml
[tool.poetry]
name = "video-resizer"
version = "0.1.0"
description = ""
authors = ["Vahid Al <thevahidal@gmail.com>"]
readme = "README.md"
packages = [{include = "video_resizer"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.3"
python-decouple = "^3.6"
django-extensions = "^3.2.1"
drf-yasg = "^1.21.4"
gunicorn = "^20.1.0"
psycopg2 = "^2.9.5"
celery = "^5.2.7"
redis = "^4.3.4"


[tool.poetry.group.dev.dependencies]
black = {version = "^22.10.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
